Add role check for calendar actions
